修改默认清晰度为 1080P60帧
4K的话要进行两次请求,没必要

+++ BiliDuang/JSONCallback/BiliPlus/AV.cs
@@ -147,7 +147,7 @@ namespace BiliDuang.JSONCallback.BiliPlus
         /// <summary>
         /// 
         /// </summary>
-        public int vipDueDate { get; set; }
+        public long vipDueDate { get; set; }
         /// <summary>
         /// 
         /// </summary>
